python/ubuntu-system-metrics/requirements.txt
psutil
influxdb-client
pyyaml
